embedded node modules and added test fixtures
I don't know if that is a good approach or not
