--- gaphor-autolayout/__init__.py
@@ -1,6 +1,0 @@
-# ruff: noqa: F401
-
-from gaphor.plugins.autolayoutELK.autolayoutelk import (
-    AutoLayoutELK,
-    AutoLayoutELKService,
-)
