+++ gaphor_autolayout/__init__.py
@@ -1,0 +1,6 @@
+# ruff: noqa: F401
+
+# from autolayoutelk import (
+#     AutoLayoutELK,
+#     AutoLayoutELKService,
+# )
