+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "elkjs": "^0.9.3"
+  }
+}
